Fix return type of getStatus in javascore/bmc and bmv to Map
According to IIP-25, return type of BMC.getStatus contains status of BMV as object.
but the JavaEE does not allow that Object or Map in struct,
so in BMC.getStatus, using Map instead of BMCStatus for containing status of any types of BMV.
also in BMV.getStatus, using Map instead of BMVStatus.